+++ app/src/main/java/apps/czeidler/economylogboook/data/FuelUnits.java
@@ -8,12 +8,12 @@ import android.preference.PreferenceManager;
  * Created by Colin on 2016-03-01.
  */
 public enum FuelUnits {
-    LITERS (1f, "L"),
-    GALLONS(3.78f, "G");
+    LITERS (1, "L"),
+    GALLONS(3.78, "G");
 
-    private final float ratio;
+    private final double ratio;
     private final String unit;
-    FuelUnits(float ratio, String unit) {
+    FuelUnits(double ratio, String unit) {
         this.ratio = ratio;
         this.unit = unit;
     }
@@ -22,7 +22,7 @@ public enum FuelUnits {
         return unit;
     }
 
-    public float getRatio() {
+    public double getRatio() {
         return ratio;
     }
 
@@ -32,7 +32,7 @@ public enum FuelUnits {
         return valueOf(fuelPref).getUnit();
     }
 
-    public static float getSystemRatio(Context c) {
+    public static double getSystemRatio(Context c) {
         SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(c);
         String fuelPref = prefs.getString("pref_fuelunit", "LITERS");
         return valueOf(fuelPref).getRatio();
